ciao-controller: Extract http server functionality into own file
Simplify main.go by extracting HTTP server functionality, including
authentication, into its own file.
