+++ ciao-controller/main.go
@@ -18,17 +18,13 @@ package main
 
 import (
 	"context"
-	"crypto/tls"
-	"crypto/x509"
 	"flag"
 	"fmt"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"os/signal"
 	"sync"
 	"syscall"
-	"time"
 
 	"github.com/01org/ciao/ciao-controller/api"
 	"github.com/01org/ciao/ciao-controller/internal/datastore"
@@ -36,13 +32,9 @@ import (
 	storage "github.com/01org/ciao/ciao-storage"
 	"github.com/01org/ciao/clogger/gloginterface"
 	"github.com/01org/ciao/database"
-	osIdentity "github.com/01org/ciao/openstack/identity"
 	"github.com/01org/ciao/osprepare"
-	"github.com/01org/ciao/service"
 	"github.com/01org/ciao/ssntp"
 	"github.com/golang/glog"
-	"github.com/gorilla/mux"
-	"github.com/pkg/errors"
 )
 
 type tenantConfirmMemo struct {
@@ -251,155 +243,4 @@ func main() {
 	ctl.ds.Exit()
 	ctl.is.ds.Shutdown()
 	ctl.client.Disconnect()
-}
-
-type clientCertAuthHandler struct {
-	Next http.Handler
-}
-
-func (h *clientCertAuthHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	if len(r.TLS.VerifiedChains) != 1 {
-		http.Error(w, "Unexpected number of certificate chains presented", http.StatusUnauthorized)
-		return
-	}
-
-	certs := r.TLS.VerifiedChains[0]
-	cert := certs[0]
-	tenants := cert.Subject.Organization
-
-	privileged := false
-	if len(tenants) == 1 && tenants[0] == "admin" {
-		privileged = true
-	}
-
-	r = r.WithContext(service.SetPrivilege(r.Context(), true))
-
-	vars := mux.Vars(r)
-	tenantFromVars := vars["tenant"]
-	if !privileged {
-		tenantMatched := false
-		for i := range tenants {
-			if tenants[i] == tenantFromVars {
-				tenantMatched = true
-				break
-			}
-		}
-		if !tenantMatched {
-			http.Error(w, "Access to tenant not permitted with certificate", http.StatusUnauthorized)
-			return
-		}
-	}
-
-	r = r.WithContext(service.SetTenantID(r.Context(), tenantFromVars))
-	h.Next.ServeHTTP(w, r)
-}
-
-func (c *controller) createCiaoRoutes(r *mux.Router) error {
-	config := api.Config{URL: c.apiURL, CiaoService: c}
-
-	r = api.Routes(config, r)
-
-	// wrap each route in keystone validation.
-	validServices := []osIdentity.ValidService{
-		{ServiceType: "compute", ServiceName: "ciao"},
-		{ServiceType: "compute", ServiceName: "nova"},
-		{ServiceType: "image", ServiceName: "glance"},
-		{ServiceType: "image", ServiceName: "ciao"},
-		{ServiceType: "volume", ServiceName: "ciao"},
-		{ServiceType: "volumev2", ServiceName: "ciao"},
-		{ServiceType: "volume", ServiceName: "cinder"},
-		{ServiceType: "volumev2", ServiceName: "cinderv2"},
-	}
-
-	validAdmins := []osIdentity.ValidAdmin{
-		{Project: "service", Role: "admin"},
-		{Project: "admin", Role: "admin"},
-	}
-
-	err := r.Walk(func(route *mux.Route, router *mux.Router, ancestors []*mux.Route) error {
-		if clientCertCAPath != "" {
-			h := &clientCertAuthHandler{
-				Next: route.GetHandler(),
-			}
-			route.Handler(h)
-		} else {
-			osHandler := osIdentity.Handler{
-				Client:        c.id.scV3,
-				Next:          route.GetHandler(),
-				ValidServices: validServices,
-				ValidAdmins:   validAdmins,
-			}
-
-			route.Handler(osHandler)
-		}
-
-		return nil
-	})
-
-	return err
-}
-
-func (c *controller) createCiaoServer() (*http.Server, error) {
-	r := mux.NewRouter()
-
-	if err := c.createComputeRoutes(r); err != nil {
-		return nil, errors.Wrap(err, "Error adding compute routes")
-	}
-
-	if err := c.createImageRoutes(r); err != nil {
-		return nil, errors.Wrap(err, "Error adding image routes")
-	}
-
-	if err := c.createVolumeRoutes(r); err != nil {
-		return nil, errors.Wrap(err, "Error adding volume routes")
-	}
-
-	err := c.createCiaoRoutes(r)
-	if err != nil {
-		return nil, errors.Wrap(err, "Error adding ciao routes")
-	}
-
-	addr := fmt.Sprintf(":%d", controllerAPIPort)
-
-	server := &http.Server{
-		Handler: r,
-		Addr:    addr,
-	}
-
-	if clientCertCAPath != "" {
-		clientCertCAbytes, err := ioutil.ReadFile(clientCertCAPath)
-		if err != nil {
-			return nil, errors.Wrap(err, "Error loading client cert CA")
-		}
-		certPool := x509.NewCertPool()
-		ok := certPool.AppendCertsFromPEM(clientCertCAbytes)
-		if !ok {
-			return nil, errors.New("Error importing client auth CA to poool")
-		}
-		tlsConfig := tls.Config{
-			ClientAuth: tls.RequireAndVerifyClientCert,
-			ClientCAs:  certPool,
-		}
-		server.TLSConfig = &tlsConfig
-	}
-
-	return server, nil
-}
-
-func (c *controller) ShutdownHTTPServers() {
-	glog.Warning("Shutting down HTTP servers")
-	var wg sync.WaitGroup
-	for _, server := range c.httpServers {
-		wg.Add(1)
-		go func(server *http.Server) {
-			ctx, cancel := context.WithTimeout(context.Background(), time.Second*5)
-			defer cancel()
-			err := server.Shutdown(ctx)
-			if err != nil {
-				glog.Errorf("Error during HTTP server shutdown")
-			}
-			wg.Done()
-		}(server)
-	}
-	wg.Wait()
 }
